refactor: dimensionstructure integration tests refactor (#214)
This PR is the first of many dealing with MasterData domain entities realated integration tests refactor.
This PR moved There is a DimensionStructure domain object to a separate project called MasterData.StepDefinitions, which is a shared project between business logic and web api tests.

This PR introduces There is a parametrized DimensionStructure domain object, where the DimensionStructure object property values defined and provided by the Scenario. It is also placed in the MasterData.StepDefinitions.

There are two project which use very similar or identical step definitions in testing, but these step defs are implemented in both cases. The goal here is to find the common ground between the two project testing needs.